+++ framework/packages/abstract-app/src/lib.rs
@@ -13,6 +13,9 @@ pub use error::AppError;
 pub type AppResult<C = Empty> = Result<Response<C>, AppError>;
 mod interface;
 
+#[cfg(feature = "cli")]
+pub mod cli;
+
 use cosmwasm_std::{Empty, Response};
 #[cfg(feature = "test-utils")]
 pub mod mock {
